Feature/#5

+++ src/main/java/com/skyhorsemanpower/payment/payment/vo/PaymentDetailResponseVo.java
@@ -10,14 +10,14 @@ public class PaymentDetailResponseVo {
 	private String paymentUuid;
 	private String auctionUuid;
 	private String paymentMethod;
-	private int price;
+	private double price;
 	private String paymentNumber;
 	private PaymentStatus paymentStatus;
 	private LocalDateTime paymentAt;
 	private LocalDateTime paymentCompletionAt;
 
 	public PaymentDetailResponseVo(String paymentUuid, String auctionUuid, String paymentMethod,
-		int price, String paymentNumber,
+		double price, String paymentNumber,
 		PaymentStatus paymentStatus, LocalDateTime paymentAt, LocalDateTime paymentCompletionAt) {
 		this.paymentUuid = paymentUuid;
 		this.auctionUuid = auctionUuid;
